FC: Add support of HMC5883L for heading measurement
This implementation is designed to provide a heading measurement based
on the magnetometer readings from the HMC5883L sensor.

The sensor is initialized with the appropriate settings such as the output data rate and the measurement mode.
Then, in the Read_HMC5883L function, the sensor is read to obtain raw X, Y, and Z magnetometer values.
These values are then scaled by the calibration factors that were determined during the calibration process.

The scaled X and Y values are used to calculate the heading angle in radians using the arctangent function.
This angle is then converted to degrees and stored in the heading variable.

Signed-off-by: pkm774 <mprabhat774@gmail.com>